requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
reportlab==4.2.2
